refactor #110 : SummaryTitle 컴포넌트 분리
