+++ src/features/total-evaluation/components/overall-evaluation/overall-evaluation.styles.ts
@@ -1,7 +1,5 @@
 import { css } from '@emotion/react';
 
-import { withTheme } from '@/common/utils/with-theme';
-
 export const overallEvaluationWrapper = css`
   display: flex;
   flex-direction: column;
@@ -13,15 +11,6 @@ export const summaryWrapper = css`
   flex-direction: column;
   gap: 1.6rem;
 `;
-
-export const summaryTitle = withTheme(
-  (theme) => css`
-    display: flex;
-    align-items: center;
-    ${theme.fonts.CAPTION.CAPTION1_SB}
-    color: ${theme.colors.GRAY[300]};
-  `,
-);
 
 export const analysisWrapper = css`
   display: flex;
