+++ src/features/total-evaluation/components/summary-title/summary-title.styles.ts
@@ -1,0 +1,12 @@
+import { css } from '@emotion/react';
+
+import { withTheme } from '@/common/utils/with-theme';
+
+export const summaryTitle = withTheme(
+  (theme) => css`
+    display: flex;
+    align-items: center;
+    ${theme.fonts.CAPTION.CAPTION1_SB}
+    color: ${theme.colors.GRAY[300]};
+  `,
+);
